Minor bugfix in ACTMCSymbolicParameterSynthesis:
When ACTMCSymbolicParameterSynthesis unsuccessfully terminated in an
unexpected manner (e.g. computing reachability rewards exceeded the
iteration limit), the ACTMC parameters did not restore their original
values. Calling other computations after this worked with an ACTMC that
had parameters different to that of the currently parsed model.
Henceforth, ACTMCSymbolicParameterSynthesis works with a hard copy of
the ACTMC, making everything more robust and code simpler.